requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
motor==3.1.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==1.10.7
pymongo==4.3.3
python-multipart==0.0.9
email-validator==2.1.1
